........
  r583575 | dkulp | 2007-10-10 14:57:20 -0400 (Wed, 10 Oct 2007) | 8 lines
  
  [CXF-986, CXF-1097, CXF-1059, CXF-1099] More "large message" fixes
   * Add configurable limit to the logging interceptors (default is 100K)
   * Fix problems with WS-RM with large messages (still uses byte[], but it at least works now)
   * Change all temp files to go through FileUtils
   * Update temp file creation to go into a directory we wipe on exit via exit hook to work around memory leak in file.deleteOnExit() (see http://bugs.sun.com/bugdatabase/view_bug.do?bug_id=4872014)
   * Fix a BUNCH of places that were leaking temp files to properly close streams so that temp files will get deleted.
   * Make HTTPConduit.finalizeConfig protected
........


git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/branches/2.0.x-fixes@583584 13f79535-47bb-0310-9956-ffa450edef68

+++ common/common/src/main/java/org/apache/cxf/helpers/FileUtils.java
@@ -20,17 +20,46 @@
 package org.apache.cxf.helpers;
 
 import java.io.File;
-import java.text.DecimalFormat;
+import java.io.IOException;
 import java.util.Locale;
-import java.util.Random;
 
 public final class FileUtils {
     private static final int RETRY_SLEEP_MILLIS = 10;
-    private static Random rand = new Random(System.currentTimeMillis()
-                                            + Runtime.getRuntime().freeMemory());
-
+    private static File defaultTempDir;
+    
+    
     private FileUtils() {
         
+    }
+    
+    private static synchronized File getDefaultTempDir() {
+        if (defaultTempDir != null) {
+            return defaultTempDir;
+        }
+        String s = System.getProperty(FileUtils.class.getName() + ".TempDirectory");
+        if (s == null) {
+            int x = (int)(Math.random() * 1000000);
+            s = System.getProperty("java.io.tmpdir");
+            File f = new File(s, "cxf-tmp-" + x);
+            while (!f.mkdir()) {
+                x = (int)(Math.random() * 1000000);
+                f = new File(s, "cxf-tmp-" + x);
+            }
+            defaultTempDir = f;
+            Thread hook = new Thread() {
+                @Override
+                public void run() {
+                    removeDir(defaultTempDir);
+                }
+            };
+            Runtime.getRuntime().addShutdownHook(hook);            
+        } else {
+            //assume someone outside of us will manage the directory
+            File f = new File(s);
+            f.mkdirs();
+            defaultTempDir = f;
+        }
+        return defaultTempDir;
     }
 
     public static void mkDir(File dir) {
@@ -108,26 +137,31 @@ public final class FileUtils {
         return osName.indexOf("windows") > -1;
     }
 
-    public static File createTempFile(String prefix, String suffix) {
-        return createTempFile(prefix, suffix, null, true);
+    public static File createTempFile(String prefix, String suffix) throws IOException {
+        return createTempFile(prefix, suffix, null, false);
     }
-
+    
     public static File createTempFile(String prefix, String suffix, File parentDir,
-                               boolean deleteOnExit) {
+                               boolean deleteOnExit) throws IOException {
         File result = null;
-        String parent = (parentDir == null)
-            ? System.getProperty("java.io.tmpdir")
-            : parentDir.getPath();
-
-        DecimalFormat fmt = new DecimalFormat("#####");
-        synchronized (rand) {
-            do {
-                result = new File(parent,
-                                  prefix + fmt.format(Math.abs(rand.nextInt()))
-                                  + suffix);
-            } while (result.exists());
+        File parent = (parentDir == null)
+            ? getDefaultTempDir()
+            : parentDir;
+            
+        if (suffix == null) {
+            suffix = ".tmp";
         }
-        if (deleteOnExit) {
+        if (prefix == null) {
+            prefix = "cxf";
+        } else if (prefix.length() < 3) {
+            prefix = prefix + "cxf";
+        }
+        result = File.createTempFile(prefix, suffix, parent);
+
+        //if parentDir is null, we're in our default dir
+        //which will get completely wiped on exit from our exit
+        //hook.  No need to set deleteOnExit() which leaks memory.
+        if (deleteOnExit && parentDir != null) {
             result.deleteOnExit();
         }
         return result;
